+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Chore & Maintenance
 
+- `[*]` Replace `any`s with `unknown`s ([#9626](https://github.com/facebook/jest/pull/9626))
 - `[@jest/transform]` Expose type `CacheKeyOptions` for `getCacheKey` ([#9762](https://github.com/facebook/jest/pull/9762))
 
 ### Performance